reuse compositional field matrix memory
This patch lets all compositional fields share the same system matrix.
As we assemble the matrices before each solve anyways, this doesn't cost
any extra time. Sadly, handling of constraints is more complicated as
the value of the boundary conditions can be different for each field.
For this, we now have separate constraint matrices for each field.